Split backend_gtk3cairo.py into two parts, one dependent on Gcf and one not.
Same naming scheme as rest.
